+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Seperate table images with users, books
 - Add POSTMAN.json
 - Add Transition to services that write to DB for ensure all operations success before write DB.
+- Add Book Management Services
 ### Noted
 - Get All book with params (Filter, Search) not implement yet need advice